Harden: fix Actions template syntax, per-case prompt-render failures, judge concurrency & cost accounting, pooled HTTP client, fixture-aware config_hash, judge-injection fencing, own CI gate workflow
- eval-gate.yml: missing $ in ${{ secrets }} would break user workflows
- runner: bad prompt_template/vars now fail the single case, not the whole run
- metrics: judge calls throttled by their own semaphore; cost/latency metered
  into MetricResult and rolled up into CaseEval.total_cost_usd / summaries
- openai provider: one pooled AsyncClient per run (was: handshake per call)
- config_hash: covers mock fixture contents, matching the cache-salt semantics
- judge prompts: <untrusted> fencing + ignore-instructions directive
- storage: pass_rate column Float (was Integer - would truncate on Postgres)
- CI: pytest + gate must pass healthy suite and FAIL regressed suite
- tests: 33 passing (render degradation, fixture hash, judge cost, html, storage)

+++ src/evalhive/storage/__init__.py
@@ -7,7 +7,7 @@ import os
 from datetime import datetime, timezone
 from pathlib import Path
 
-from sqlalchemy import JSON, DateTime, Integer, String, Text, create_engine, select
+from sqlalchemy import JSON, DateTime, Float, Integer, String, Text, create_engine, select
 from sqlalchemy.orm import DeclarativeBase, Mapped, Session, mapped_column
 
 from ..core.results import RunResult
@@ -30,7 +30,7 @@ class RunRow(Base):
     error: Mapped[str | None] = mapped_column(Text, nullable=True)
     n_cases: Mapped[int] = mapped_column(Integer, default=0)
     n_passed: Mapped[int] = mapped_column(Integer, default=0)
-    pass_rate: Mapped[float] = mapped_column(Integer, default=0.0)
+    pass_rate: Mapped[float] = mapped_column(Float, default=0.0)
     payload: Mapped[dict | None] = mapped_column(JSON, nullable=True)
     is_baseline: Mapped[int] = mapped_column(Integer, default=0)
